Reorganized the directory structure
