+++ LICENSE.txt
@@ -1,4 +1,6 @@
-Copyright (c) 2012 Kenneth Falck <kennu@iki.fi> http://kfalck.net
+Copyright (c) 2013 Merchant Atlas Inc. http://www.merchantatlas.com
+
+[Original Work] Copyright (c) 2012 Kenneth Falck <kennu@iki.fi> http://kfalck.net
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
