+++ setup.py
@@ -2,11 +2,11 @@ from distutils.core import setup
 
 setup(
     name='django-mongom2m',
-    version='0.1.0',
-    author=u'Kenneth Falck',
-    author_email='kennu@iki.fi',
-    packages=['mongom2m', 'mongom2m_testapp'],
-    url='https://github.com/kennu/django-mongom2m',
+    version='0.2.0',
+    author=u'Merchant Atlas Inc.',
+    author_email='support@merchantatlas.com',
+    packages=['django_mongom2m'],
+    url='https://github.com/mobilespinach/django-mongom2m',
     license='BSD licence, see LICENCE.txt',
     description='A ManyToManyField for django-mongodb-engine',
     long_description='A ManyToManyField for django-mongodb-engine',
